chat/show/UI調整

+++ resources/views/chats/index.blade.php
@@ -22,13 +22,16 @@
     &nbsp;
     <br />
     <div class="panel-body">
+        <?php // var_dump($user->id ); ?>
         <table class="table table-striped task-table">
             <thead>
                 <th>Id</th>
                 <th>Name</th>
-                <th>Action</th>
-                <th>Edit</th>
+                <th>Join</th>
+                <th>action</th>
+                <!--
                 <th>Delete</th>
+                -->
             </thead>
             <?php //debug_dump($tasks); ?>
             <tbody>
@@ -37,11 +40,13 @@
                         <td class="table-text">{{$chat->id}}
                         </td>
                         <td class="table-text">
-                            <h3>{{ link_to_route('chats.show', $chat->name, $chat->id) }}
+                            <h3>{{ link_to_route('chats.show', $chat->name, $chat->id, 
+                                ) }}
                             </h3>
                         </td>
                         <td class="table-text">
-                            <a href="chats/add_member?cid={{$chat->id}}">[ 参加する ]
+                            <a href="chats/add_member?cid={{$chat->id}}"
+                                 class="btn btn-outline-primary btn-sm">参加する
                             </a>
                             <!--
                             <a href="chats/delete_member?cid={{$chat->id}}">[ 退会 ]
@@ -49,14 +54,19 @@
                             -->
                         </td>
                         <td class="table-text">
-                            {{ link_to_route('chats.edit', '編集'
-                            , $chat->id, ['class' => 'btn btn-sm btn-default']) }}
-                        </td>
-                        <td class="table-text">
-                            {{ Form::open(['route' => ['chats.destroy', $chat->id], 'method' => 'delete']) }}
+                            <div style="float :left; margin-right :10px">
+                                <?php if($user->id == $chat->user_id){ ?>
+                                    {{ link_to_route('chats.edit', '編集'
+                                    , $chat->id, ['class' => 'btn btn-outline-primary btn-sm']) }}                                
+                                <?php } ?>
+                            </div>
+                            <?php if($user->id == $chat->user_id){ ?>
+                                {{ Form::open(['route' => ['chats.destroy', $chat->id], 'method' => 'delete']) }}
                                 {{ Form::hidden('id', $chat->id) }}
-                                {{ Form::submit('削除', ['class' => 'btn btn-sm btn-default']) }}
-                            {{ Form::close() }}
+                                {{ Form::submit('削除', ['class' => 'btn btn-outline-danger btn-sm']) }}
+                                {{ Form::close() }}                                
+                            <?php } ?>                            
+
                         </td>
                     </tr>
                 @endforeach
